- Support for VS2015 RC  - Support for LazurBeemz 6.1  - Some minor syntax updates (mostly brace-constructor style and member-brace-init style)  - Fixing some compile errors on PacketReader::skipBytes  - Fixing up some compile warnings that came with VS2015 (size_t used to implicitly cast to smaller integral types, this is no longer the case)  - Adding overloads to Timer::Id so downstream doesn't have to care about the signedness of the integers  - Fixing up some compile errors that came with VS2015 (stricter interpretation on assignment of time_point_t and duration_t values)  - Modifying some copies to assignment now that assignment is not being automatically generated in cases where the constructor is specified  - Removing difference between Boost's v3 filesystem API and the out of the box implementation for Visual Studio (they use the same API now)  - Removing thread_local hack for MSVC  - Removing discrete_distribution ugly hack for MSVC  - Removing reverse_iterator declarations from ObjectPool, std::unordered_map no longer has reverse_iterator
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/buis_vana_overhaul@3474 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/InventoryPacket.cpp
@@ -134,7 +134,7 @@ PACKET_IMPL(inventoryOperation, bool unk, const vector_t<InventoryPacketOperatio
 	builder
 		.add<header_t>(SMSG_INVENTORY_OPERATION)
 		.add<bool>(unk)
-		.add<uint8_t>(operations.size());
+		.add<uint8_t>(static_cast<uint8_t>(operations.size()));
 
 	int8_t addedByte = -1;
 
@@ -382,7 +382,7 @@ PACKET_IMPL(sendItemExpired, const vector_t<item_id_t> &items) {
 	builder
 		.add<int16_t>(SMSG_NOTICE)
 		.add<int8_t>(0x08)
-		.add<uint8_t>(items.size());
+		.add<uint8_t>(static_cast<uint8_t>(items.size()));
 
 	for (const auto &itemId : items) {
 		builder.add<item_id_t>(itemId);
